% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{computePreferenceScore}
\alias{computePreferenceScore}
\title{Compute Preference Score}
\usage{
computePreferenceScore(data, unfilteredData = NULL)
}
\arguments{
\item{data}{A data frame containing propensity scores.}

\item{unfilteredData}{Optional dataset to compute proportions from.}
}
\value{
A data frame with added preference scores.
}
\description{
Computes preference scores based on the propensity scores.
}
