% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dTTE.R
\name{plot.dTTE}
\alias{plot.dTTE}
\title{Plot Method for dTTE Objects (Calibration Only)}
\usage{
\method{plot}{dTTE}(x, ...)
}
\arguments{
\item{x}{An object of class "dTTE".}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
Plots only the calibration graph for the dTTE pipeline output.
}
