% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dTTE.R
\name{plot.TTE}
\alias{plot.TTE}
\title{Plot Method for TTE Objects}
\usage{
\method{plot}{TTE}(x, which = c("SMD", "Equipoise"), ...)
}
\arguments{
\item{x}{An object of class "TTE".}

\item{which}{A character vector specifying which plot(s) to display.
Options are "SMD" (Standardized Mean Differences) and
"Equipoise" (Equipoise plot). Default shows all plots.}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
Plots diagnostic graphs for the TTE pipeline output.
}
