% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dTTE.R
\name{print.dTTE}
\alias{print.dTTE}
\title{Print Method for dTTE Objects}
\usage{
\method{print}{dTTE}(x, ...)
}
\arguments{
\item{x}{An object of class "dTTE".}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Prints a concise summary of the dTTE pipeline output.
}
\value{
  No return value, printed to the console.
}
