% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dTTE.R
\name{print.TTE}
\alias{print.TTE}
\title{Print Method for TTE Objects}
\usage{
\method{print}{TTE}(x, ...)
}
\arguments{
\item{x}{An object of class "TTE".}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Prints a concise summary of the TTE pipeline output.
}
