% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dTTE.R
\name{summary.dTTE}
\alias{summary.dTTE}
\title{Summary Method for dTTE Objects}
\usage{
\method{summary}{dTTE}(x, ...)
}
\arguments{
\item{x}{An object of class "dTTE".}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Provides a detailed summary of the dTTE pipeline output.
}
