% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dTTE.R
\name{summary.TTE}
\alias{summary.TTE}
\title{Summary Method for TTE Objects}
\usage{
\method{summary}{TTE}(x, ...)
}
\arguments{
\item{x}{An object of class "TTE".}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Provides a detailed summary of the TTE pipeline output.
}
\value{
  No return value, printed to the console.
}
